refactoring DashboardPage: delete locators_dashboard_page.py; moved all locators into DashboardPage.py; added descriptions into BasePage.py, DashboardPage.py and test_dashboard_page.py
